+++ src/main/java/net/thedragonteam/armorplus/blocks/base/BaseBenchBlock.java
@@ -6,10 +6,17 @@ package net.thedragonteam.armorplus.blocks.base;
 
 import net.minecraft.block.ITileEntityProvider;
 import net.minecraft.block.material.Material;
+import net.minecraft.block.properties.PropertyDirection;
+import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumBlockRenderType;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.Rotation;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -17,10 +24,13 @@ import net.thedragonteam.armorplus.tileentity.base.TileEntityBaseBench;
 
 public class BaseBenchBlock extends BaseBlock implements ITileEntityProvider {
 
+    public static final PropertyDirection FACING = PropertyDirection.create("facing", EnumFacing.Plane.HORIZONTAL);
+
     public TileEntityBaseBench tileEntityBaseBench;
 
     public BaseBenchBlock(String name, TileEntityBaseBench tileEntityBaseBench) {
         super(Material.IRON, name, 1000.0F, 10.0F, "pickaxe", 2);
+        setDefaultState(this.blockState.getBaseState().withProperty(FACING, EnumFacing.NORTH));
         this.tileEntityBaseBench = tileEntityBaseBench;
     }
 
@@ -49,4 +59,35 @@ public class BaseBenchBlock extends BaseBlock implements ITileEntityProvider {
     public TileEntity createNewTileEntity(World worldIn, int meta) {
         return tileEntityBaseBench;
     }
+
+
+    @Override
+    public IBlockState getStateForPlacement(World worldIn, BlockPos pos, EnumFacing facing, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer, ItemStack stack) {
+        IBlockState iblockstate = super.onBlockPlaced(worldIn, pos, facing, hitX, hitY, hitZ, meta, placer);
+        iblockstate = iblockstate.withProperty(FACING, placer.getHorizontalFacing());
+        return iblockstate;
+    }
+
+    @Override
+    public int getMetaFromState(IBlockState state) {
+        return state.getValue(FACING).getHorizontalIndex();
+    }
+
+    @Override
+    public IBlockState getStateFromMeta(int meta) {
+        IBlockState iblockstate = this.getDefaultState();
+        iblockstate = iblockstate.withProperty(FACING, EnumFacing.getHorizontal(meta));
+        return iblockstate;
+    }
+
+    @Override
+    public IBlockState withRotation(IBlockState state, Rotation rot) {
+        return state.withProperty(FACING, rot.rotate((EnumFacing) state.getValue(FACING)));
+    }
+
+    @Override
+    protected BlockStateContainer createBlockState() {
+        return new BlockStateContainer(this, FACING);
+    }
+
 }