+++ src/main/java/net/thedragonteam/armorplus/compat/JEI.java
@@ -25,7 +25,6 @@ public class JEI implements IModPlugin {
         blacklist.addItemToBlacklist(new ItemStack(ModItems.theDragonTeamItem, 1));
         blacklist.addItemToBlacklist(new ItemStack(ModItems.twitchItem, 1));
         blacklist.addItemToBlacklist(new ItemStack(ModItems.beamItem, 1));
-        blacklist.addItemToBlacklist(new ItemStack(ModItems.mmdItem, 1));
         blacklist.addItemToBlacklist(new ItemStack(ModItems.nbtItem, 1));
         blacklist.addItemToBlacklist(new ItemStack(ModItems.devHelmet, 1));
         blacklist.addItemToBlacklist(new ItemStack(ModItems.devChestplate, 1));