+++ src/main/java/net/thedragonteam/armorplus/items/base/energy/rf/BaseRFPickaxe.java
@@ -9,6 +9,7 @@ import net.minecraft.block.Block;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -16,6 +17,8 @@ import net.thedragonteam.armorplus.ArmorPlus;
 import net.thedragonteam.armorplus.items.base.BasePickaxe;
 
 import java.util.Set;
+
+import static net.thedragonteam.thedragonlib.util.TextHelper.localize;
 
 public class BaseRFPickaxe extends BasePickaxe implements IEnergyContainerItem {
 
@@ -37,6 +40,11 @@ public class BaseRFPickaxe extends BasePickaxe implements IEnergyContainerItem {
         this.capacity = capacity;
         this.maxReceive = maxReceive;
         this.maxExtract = maxExtract;
+    }
+
+    @Override
+    public String getItemStackDisplayName(ItemStack stack) {
+        return (TextFormatting.DARK_RED + localize(this.getUnlocalizedNameInefficiently(stack) + ".name")).trim();
     }
 
     @Override