+++ src/main/java/net/thedragonteam/armorplus/items/dev/NBTItem.java
@@ -4,7 +4,6 @@
 
 package net.thedragonteam.armorplus.items.dev;
 
-import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
@@ -12,21 +11,14 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
-import net.minecraftforge.client.model.ModelLoader;
-import net.minecraftforge.fml.common.registry.GameRegistry;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-import net.thedragonteam.armorplus.ArmorPlus;
+import net.thedragonteam.armorplus.items.base.BaseItem;
 
 import java.util.List;
 
-public class NBTItem extends Item {
+public class NBTItem extends BaseItem {
 
     public NBTItem() {
-        setRegistryName("dev_item");
-        setUnlocalizedName(ArmorPlus.MODID + "." + "dev_item");
-        GameRegistry.register(this);
-        this.setCreativeTab(ArmorPlus.tabArmorplusItems);
+        super("dev_item");
     }
 
     @Override
@@ -67,11 +59,6 @@ public class NBTItem extends Item {
     @Override
     public Item setCreativeTab(CreativeTabs tab) {
         return null;
-    }
-
-    @SideOnly(Side.CLIENT)
-    public void initModel() {
-        ModelLoader.setCustomModelResourceLocation(this, 0, new ModelResourceLocation(getRegistryName(), "inventory"));
     }
 
     @Override