+++ src/main/java/net/thedragonteam/armorplus/util/Utils.java
@@ -9,16 +9,9 @@ import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.client.renderer.block.model.ItemCameraTransforms;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.fml.common.discovery.ASMDataTable;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import net.thedragonteam.armorplus.ArmorPlus;
 import net.thedragonteam.armorplus.api.util.NBTHelper;
-
-import java.util.Random;
-import java.util.Set;
 
 /**
  * net.thedragonteam.armorplus.util
@@ -27,31 +20,11 @@ import java.util.Set;
  */
 public final class Utils {
 
-    public static final String RESOURCE = ArmorPlus.MODID;
-
-    public static final Random RANDOM = new Random();
-
-    public static ResourceLocation getResource(String res) {
-        return new ResourceLocation(RESOURCE, res);
-    }
-
     @SuppressWarnings("ConstantConditions")
     public static ItemStack setUnbreakable(ItemStack stack) {
         NBTHelper.checkNBT(stack);
         stack.getTagCompound().setBoolean("Unbreakable", true);
         return stack;
-    }
-
-    public static void registerHandlers(Set<ASMDataTable.ASMData> eventHandlers) {
-        for (ASMDataTable.ASMData data : eventHandlers) {
-            try {
-                Class<?> handlerClass = Class.forName(data.getClassName());
-                Object handlerImpl = handlerClass.newInstance();
-                MinecraftForge.EVENT_BUS.register(handlerImpl);
-            } catch (Exception e) {
-                // No-op
-            }
-        }
     }
 
     @SideOnly(Side.CLIENT)